Refactor CirclePage to streamline login and dashboard components
- Removed unused globe rendering and related configurations for a cleaner codebase.
- Integrated CircleLogin and CircleDashboard components for improved structure and maintainability.
- Simplified state management by eliminating unnecessary variables and effects.
